backend/requirements.txt
aiocache==0.11.1
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.24.0
atomicwrites==1.4.0
attrs==21.4.0
Authlib==1.0.1
bcrypt==3.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
coverage==6.3.2
cryptography==3.4.8
dependency-injector==4.36.0
fastapi==0.68.1
fastapi-pagination==0.9.1
google-api-core==2.7.3
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.5.1
google-cloud==0.34.0
google-cloud-dialogflow==2.13.0
googleapis-common-protos==1.53.0
googlemaps==4.5.3
greenlet==1.1.2
grpcio==1.46.1
grpcio-status==1.46.1
h11==0.13.0
httpcore==0.14.7
httplib2==0.20.4
httpx==0.21.1
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Mako==1.2.0
MarkupSafe==2.1.1
motor==2.5.1
msal==1.14.0
oauth2client==4.1.3
oauthlib==3.1.1
overrides==6.1.0
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
proto-plus==1.20.3
protobuf==3.20.1
psycopg2==2.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
PyJWT==2.1.0
pymongo==3.12.3
pyparsing==3.0.8
pytest==6.2.5
pytest-cov==2.12.1
pytest-html==3.1.1
pytest-metadata==2.0.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.23
starlette==0.14.2
toml==0.10.2
typing-utils==0.1.0
typing_extensions==4.2.0
uritemplate==3.0.1
urllib3==1.26.9
uvicorn==0.15.0
